data_archiver/requirements.txt
kafka-python
influxdb
influxdb_client
pymongo
numpy
obspy